plugins/oidcplugin/requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=base.txt base.in
#
# Install mozilla-django-oidc for OIDC authentication
mozilla-django-oidc==8.1.0